Add Menu Feature to Search by Entry #.
As part of the assignment, we added a menu option to search the database
by entry number. The program now accepts a number, checks it to make sure
it's an integer, and returns it. Code to actually retrieve the entry
was not written, as it is not part of the assignment.